Qualification test only checks previous sections
The reason for this is that every time your clear a section, you will
instantly be failing the test until you get the two tetrises to fulfil
the requirements again. During play, it would be difficult for the player
to check whether they still have the requirements completed when near a
section boundary.